feat: docs panel
